backend/pyproject.toml
[project]
name = "todo_app_backend"
version = "0.1.0"
description = "this is simple backend for todo_app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.135.1",
    "passlib[bcrypt]==1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.34",
    "uvicorn>=0.41.0",
]
